refactor: move CommandStatus from maestro-cli to maestro-orchestra
Relocates the enum to maestro.orchestra.debug so cloud (maestro-worker)
can reference it once the shared TestOutputWriter lands. CLI call sites
updated to import from the new FQN; no behavior change.
